Add files via upload
The modules __main__.py, build_complex.py and execute.py were performing many of the same tasks, so I reorganised the code to avoid wasting memory space. The module execute.py is now obsolete. The module __main__.py must be run from a directory outside the dccstructure package and outputs the information pertinent to Elijah's kinetics code (incidence and adjacency matrices, 2-cell normals, etc). The module build_complex.py must be run from the directory of the dccstructure package. Before the present update, it defined a function build_complex() which has now been moved to the module build.py so that it is self-contained. The module build_complex.py simply calls this function and outputs the node coordinates and the edges, faces and volumes as lists of the constituent nodes.